+++ resources/postgres_migrations/003-add-grimoire-style-tables.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE versions
 (
-    id          bigint primary key,
+    id          serial primary key,
     name        text not null,
     group_id    text not null,
     artifact_id text not null,
@@ -10,23 +10,23 @@ CREATE TABLE versions
 --;;
 CREATE TABLE namespaces
 (
-    id         bigint not null primary key,
+    id         serial not null primary key,
     name       text   not null,
     platform   text   not null,
     meta       bytea  not null,
-    version_id bigint not null,
+    version_id serial not null,
     foreign key (version_id) references versions (id),
     unique (version_id, platform, name)
 );
 --;;
 CREATE TABLE vars
 (
-    id         bigint not null primary key,
+    id         serial not null primary key,
     name       text   not null,
     platform   text   not null,
     namespace  text   not null,
     meta       bytea  not null,
-    version_id bigint not null,
+    version_id serial not null,
     foreign key (version_id) references versions (id),
     unique (version_id, platform, namespace, name)
 );
